backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
chromadb==0.5.0
langchain==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.0
google-generativeai==0.8.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
PyPDF2==3.0.1
httpx==0.27.0
